requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.5
django-admin-sortable2==2.1.9
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
Faker==19.4.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
phonenumberslite==8.13.19
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
uritemplate==4.1.1
